// src/app/api/model/getObjectTypeSchema200Response.ts
/**
 * Metasys REST API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ViewsInner } from './viewsInner';


export interface GetObjectTypeSchema200Response { 
    /**
     * A JSON schema that describes valid payloads for creating an object of the specified type.
     */
    schema?: object;
    /**
     * A nested collection of logical property groupings used for display purposes only.
     */
    views?: Array<ViewsInner>;
}